Flytt funksjonell kode fra service til utils
