+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/steg/VilkårsvurderingForNyBehandlingService.kt
@@ -2,32 +2,21 @@ package no.nav.familie.ba.sak.kjerne.steg
 
 import no.nav.familie.ba.sak.common.Feil
 import no.nav.familie.ba.sak.common.FunksjonellFeil
-import no.nav.familie.ba.sak.common.til18ÅrsVilkårsdato
-import no.nav.familie.ba.sak.kjerne.autovedtak.fødselshendelse.Resultat
 import no.nav.familie.ba.sak.kjerne.behandling.BehandlingService
 import no.nav.familie.ba.sak.kjerne.behandling.behandlingstema.BehandlingstemaService
 import no.nav.familie.ba.sak.kjerne.behandling.domene.Behandling
-import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingType
-import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingUnderkategori
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingÅrsak
-import no.nav.familie.ba.sak.kjerne.beregning.domene.YtelseType
 import no.nav.familie.ba.sak.kjerne.endretutbetaling.EndretUtbetalingAndelService
-import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.Person
 import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersonType
 import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersongrunnlagService
-import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersonopplysningGrunnlag
-import no.nav.familie.ba.sak.kjerne.personident.Aktør
 import no.nav.familie.ba.sak.kjerne.personident.PersonidentService
+import no.nav.familie.ba.sak.kjerne.steg.VilkårsvurderingForNyBehandlingUtils.genererInitiellVilkårsvurdering
+import no.nav.familie.ba.sak.kjerne.steg.VilkårsvurderingForNyBehandlingUtils.genererVilkårsvurderingFraForrigeVedtattBehandling
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårService
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingMetrics
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingService
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingUtils
-import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.PersonResultat
-import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.Vilkår
-import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.VilkårResultat
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.Vilkårsvurdering
-import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.genererPersonResultatForPerson
-import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.gjelderAlltidFraBarnetsFødselsdato
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
 import java.time.LocalDate
@@ -192,205 +181,6 @@ class VilkårsvurderingForNyBehandlingService(
         } else {
             vilkårsvurderingService.lagreInitielt(initiellVilkårsvurdering)
         }
-    }
-
-    fun genererInitiellVilkårsvurdering(
-        behandling: Behandling,
-        barnaAktørSomAlleredeErVurdert: List<Aktør>,
-        personopplysningGrunnlag: PersonopplysningGrunnlag
-    ): Vilkårsvurdering {
-        return Vilkårsvurdering(behandling = behandling).apply {
-            when {
-                behandling.type == BehandlingType.MIGRERING_FRA_INFOTRYGD &&
-                    behandling.opprettetÅrsak == BehandlingÅrsak.MIGRERING -> {
-                    personResultater = lagVilkårsvurderingForMigreringsbehandling(
-                        vilkårsvurdering = this,
-                        personopplysningGrunnlag = personopplysningGrunnlag
-                    )
-                }
-                behandling.opprettetÅrsak == BehandlingÅrsak.FØDSELSHENDELSE -> {
-                    personResultater = lagVilkårsvurderingForFødselshendelse(
-                        vilkårsvurdering = this,
-                        barnaAktørSomAlleredeErVurdert = barnaAktørSomAlleredeErVurdert,
-                        personopplysningGrunnlag = personopplysningGrunnlag
-                    )
-                }
-                !behandling.skalBehandlesAutomatisk -> {
-                    personResultater = lagManuellVilkårsvurdering(
-                        vilkårsvurdering = this,
-                        personopplysningGrunnlag = personopplysningGrunnlag
-                    )
-                }
-                else -> personResultater = lagTomVilkårsvurdering(
-                    vilkårsvurdering = this,
-                    personopplysningGrunnlag = personopplysningGrunnlag
-                )
-            }
-        }
-    }
-
-    private fun genererVilkårsvurderingFraForrigeVedtattBehandling(
-        initiellVilkårsvurdering: Vilkårsvurdering,
-        forrigeBehandlingVilkårsvurdering: Vilkårsvurdering,
-        behandling: Behandling,
-        personopplysningGrunnlag: PersonopplysningGrunnlag,
-        løpendeUnderkategori: BehandlingUnderkategori?
-    ): Vilkårsvurdering {
-        val (vilkårsvurdering) = VilkårsvurderingUtils.flyttResultaterTilInitielt(
-            aktivVilkårsvurdering = forrigeBehandlingVilkårsvurdering,
-            initiellVilkårsvurdering = initiellVilkårsvurdering,
-            løpendeUnderkategori = løpendeUnderkategori,
-            forrigeBehandlingVilkårsvurdering = forrigeBehandlingVilkårsvurdering
-        )
-
-        if (behandling.type == BehandlingType.REVURDERING && behandling.opprettetÅrsak == BehandlingÅrsak.DØDSFALL_BRUKER) {
-            vilkårsvurdering.personResultater.single { it.erSøkersResultater() }.vilkårResultater.forEach { vilkårResultat ->
-                vilkårResultat.periodeTom = personopplysningGrunnlag.søker.dødsfall?.dødsfallDato
-            }
-        }
-        return vilkårsvurdering
-    }
-
-    private fun lagVilkårsvurderingForMigreringsbehandling(
-        vilkårsvurdering: Vilkårsvurdering,
-        personopplysningGrunnlag: PersonopplysningGrunnlag
-    ): Set<PersonResultat> {
-        return personopplysningGrunnlag.søkerOgBarn.map { person ->
-            val personResultat = PersonResultat(vilkårsvurdering = vilkårsvurdering, aktør = person.aktør)
-
-            // NB Dette må gjøres om når vi skal begynne å migrere EØS-saker
-            val ytelseType = if (person.type == PersonType.SØKER) when (vilkårsvurdering.behandling.underkategori) {
-                BehandlingUnderkategori.UTVIDET -> YtelseType.UTVIDET_BARNETRYGD
-                BehandlingUnderkategori.ORDINÆR -> YtelseType.ORDINÆR_BARNETRYGD
-            } else YtelseType.ORDINÆR_BARNETRYGD
-
-            val vilkårTyperForPerson = Vilkår.hentVilkårFor(person.type, ytelseType = ytelseType)
-
-            val vilkårResultater = vilkårTyperForPerson.map { vilkår ->
-                val fom = if (vilkår.gjelderAlltidFraBarnetsFødselsdato()) person.fødselsdato else null
-
-                val tom: LocalDate? =
-                    if (vilkår == Vilkår.UNDER_18_ÅR) person.fødselsdato.plusYears(18).minusDays(1) else null
-
-                val begrunnelse = "Migrering"
-
-                VilkårResultat(
-                    personResultat = personResultat,
-                    erAutomatiskVurdert = false,
-                    resultat = Resultat.OPPFYLT,
-                    vilkårType = vilkår,
-                    periodeFom = fom,
-                    periodeTom = tom,
-                    begrunnelse = begrunnelse,
-                    behandlingId = personResultat.vilkårsvurdering.behandling.id
-                )
-            }.toSortedSet(VilkårResultat.VilkårResultatComparator)
-
-            personResultat.setSortedVilkårResultater(vilkårResultater)
-
-            personResultat
-        }.toSet()
-    }
-
-    private fun lagVilkårsvurderingForFødselshendelse(
-        vilkårsvurdering: Vilkårsvurdering,
-        barnaAktørSomAlleredeErVurdert: List<Aktør>,
-        personopplysningGrunnlag: PersonopplysningGrunnlag
-    ): Set<PersonResultat> {
-        val annenForelder = personopplysningGrunnlag.annenForelder
-        val eldsteBarnSomVurderesSinFødselsdato =
-            personopplysningGrunnlag.barna.filter { !barnaAktørSomAlleredeErVurdert.contains(it.aktør) }
-                .maxByOrNull { it.fødselsdato }?.fødselsdato
-                ?: throw Feil("Finner ingen barn på persongrunnlag")
-
-        return personopplysningGrunnlag.søkerOgBarn.map { person ->
-            val personResultat = PersonResultat(vilkårsvurdering = vilkårsvurdering, aktør = person.aktør)
-
-            val vilkårForPerson = Vilkår.hentVilkårFor(person.type)
-
-            val vilkårResultater = vilkårForPerson.map { vilkår ->
-                genererVilkårResultatForEtVilkårPåEnPerson(
-                    person = person,
-                    annenForelder = annenForelder,
-                    eldsteBarnSinFødselsdato = eldsteBarnSomVurderesSinFødselsdato,
-                    personResultat = personResultat,
-                    vilkår = vilkår
-                )
-            }
-
-            personResultat.setSortedVilkårResultater(vilkårResultater.toSet())
-
-            personResultat
-        }.toSet()
-    }
-
-    private fun lagManuellVilkårsvurdering(
-        vilkårsvurdering: Vilkårsvurdering,
-        personopplysningGrunnlag: PersonopplysningGrunnlag
-    ): Set<PersonResultat> {
-        return personopplysningGrunnlag.søkerOgBarn.map { person ->
-            genererPersonResultatForPerson(vilkårsvurdering, person)
-        }.toSet()
-    }
-
-    private fun lagTomVilkårsvurdering(
-        vilkårsvurdering: Vilkårsvurdering,
-        personopplysningGrunnlag: PersonopplysningGrunnlag
-    ): Set<PersonResultat> {
-        return personopplysningGrunnlag.søkerOgBarn.map { person ->
-            val personResultat = PersonResultat(vilkårsvurdering = vilkårsvurdering, aktør = person.aktør)
-
-            val vilkårForPerson = Vilkår.hentVilkårFor(person.type)
-
-            val vilkårResultater = vilkårForPerson.map { vilkår ->
-                VilkårResultat(
-                    personResultat = personResultat,
-                    erAutomatiskVurdert = true,
-                    resultat = Resultat.IKKE_VURDERT,
-                    vilkårType = vilkår,
-                    begrunnelse = "",
-                    behandlingId = personResultat.vilkårsvurdering.behandling.id,
-                )
-            }.toSortedSet(VilkårResultat.VilkårResultatComparator)
-
-            personResultat.setSortedVilkårResultater(vilkårResultater)
-
-            personResultat
-        }.toSet()
-    }
-
-    private fun genererVilkårResultatForEtVilkårPåEnPerson(
-        person: Person,
-        eldsteBarnSinFødselsdato: LocalDate,
-        personResultat: PersonResultat,
-        vilkår: Vilkår,
-        annenForelder: Person? = null,
-    ): VilkårResultat {
-        val automatiskVurderingResultat = vilkår.vurderVilkår(
-            person = person,
-            annenForelder = annenForelder,
-            vurderFra = eldsteBarnSinFødselsdato
-        )
-
-        val fom = if (eldsteBarnSinFødselsdato >= person.fødselsdato) eldsteBarnSinFødselsdato else person.fødselsdato
-
-        val tom: LocalDate? =
-            if (vilkår == Vilkår.UNDER_18_ÅR) {
-                person.fødselsdato.til18ÅrsVilkårsdato()
-            } else null
-
-        return VilkårResultat(
-            regelInput = automatiskVurderingResultat.regelInput,
-            personResultat = personResultat,
-            erAutomatiskVurdert = true,
-            resultat = automatiskVurderingResultat.resultat,
-            vilkårType = vilkår,
-            periodeFom = fom,
-            periodeTom = tom,
-            begrunnelse = "Vurdert og satt automatisk: ${automatiskVurderingResultat.evaluering.begrunnelse}",
-            behandlingId = personResultat.vilkårsvurdering.behandling.id,
-            evalueringÅrsaker = automatiskVurderingResultat.evaluering.evalueringÅrsaker.map { it.toString() }
-        )
     }
 
     private fun førstegangskjøringAvVilkårsvurdering(behandlingId: Long): Boolean {
