+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/vilkårsvurdering/PersonResultatUtils.kt
@@ -1,0 +1,45 @@
+package no.nav.familie.ba.sak.kjerne.vilkårsvurdering
+
+import no.nav.familie.ba.sak.common.til18ÅrsVilkårsdato
+import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.Person
+import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.PersonResultat
+import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.Vilkår
+import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.VilkårResultat
+import java.time.LocalDate
+
+object PersonResultatUtils {
+
+    fun genererVilkårResultatForEtVilkårPåEnPerson(
+        person: Person,
+        eldsteBarnSinFødselsdato: LocalDate,
+        personResultat: PersonResultat,
+        vilkår: Vilkår,
+        annenForelder: Person? = null,
+    ): VilkårResultat {
+        val automatiskVurderingResultat = vilkår.vurderVilkår(
+            person = person,
+            annenForelder = annenForelder,
+            vurderFra = eldsteBarnSinFødselsdato
+        )
+
+        val fom = if (eldsteBarnSinFødselsdato >= person.fødselsdato) eldsteBarnSinFødselsdato else person.fødselsdato
+
+        val tom: LocalDate? =
+            if (vilkår == Vilkår.UNDER_18_ÅR) {
+                person.fødselsdato.til18ÅrsVilkårsdato()
+            } else null
+
+        return VilkårResultat(
+            regelInput = automatiskVurderingResultat.regelInput,
+            personResultat = personResultat,
+            erAutomatiskVurdert = true,
+            resultat = automatiskVurderingResultat.resultat,
+            vilkårType = vilkår,
+            periodeFom = fom,
+            periodeTom = tom,
+            begrunnelse = "Vurdert og satt automatisk: ${automatiskVurderingResultat.evaluering.begrunnelse}",
+            behandlingId = personResultat.vilkårsvurdering.behandling.id,
+            evalueringÅrsaker = automatiskVurderingResultat.evaluering.evalueringÅrsaker.map { it.toString() }
+        )
+    }
+}
